Move namespace to gtnh_velo
