+++ gtnh_velo/prototypes/sympy.py
